Extensive changes to the activity grid which attempt to describe the new V2 design
After a lengthy discussion, a new 'core' design for zone was proposed.
The changes were so extensive that we christened the re-design "V2". At
a high-level, the design changes can be summarised as follows: removal
of the 'Scratchpad' and 'Contacts Drawer', to be replaced by two
'views'; "Home" (All Contacts) and Group views (multiple views, one for
each Group). This revision of the activity grid is the first shot at
re-framing the new Activities, Tasks, Actions and Operations which
users will undertake with a V2 design as the context (facilitator?).