Completed detail for the "Send a Meeting Request to my Contacts" Activity and added an "Issue List" to the Activity Grid.
Moved on and added detail to the "Send a Meeting Request to my Contacts" Activity. Also added an "Issue List" as I thought it would be handy (for now at least) to have any issues or useful info recorded in the same document as opposed to somewhere else. We can remove the issues as they're resolved and if the list becomes so big that it affects the utility of the Activity Grid, they can be moved out. Let's see how we go...