Added feedback from review, numbered headings and "Core Ideas" section.
David and I reviewed the first revision of the ISN. David had some feedback, which I made notes on during our call and have now added to the ISN (see various "Idea" bullet points).

David and I both quickly discovered during our meeting that trying to refer to any particular screen was very difficult without reference numbers and so I've added numbering to each screen title. Unfortunately Adobe Illustrator doesn't do "Numbering" so I've had to do this manually. Not ideal, but it will have to do for now.

I have also added a "Core Ideas" sidebar. The thinking here is that this section contains the main ideas or concepts of the design. These are contained in bullets on the ISN, but they're dotted around a very large space and therefore difficult to find. The "Core Ideas" sidebar is a place to bring all those ideas and thoughts together in one place so that they're easy to find and easy to read.