Modified Group Edit Screen illustrations, removing the "triangle" icon
Modified Group Edit Screen illustrations, removing the "triangle" icon because David and I agreed that it didn't make sense to be able to switch groups when in group edit mode. Doing this also makes the most sensible/intuitive gesture for editing the group name "tap" instead of the previous "tap + hold".